testsuite: Fix up anon4.C test for excess precision targets
The g++.dg/reflect/anon4.C test FAILs on i686-linux, because with
extended precision the store to .c rounds to double precision, but comparison
of it against 3.14 constant doesn't.

2026-03-02  Jakub Jelinek  <jakub@redhat.com>

	* g++.dg/reflect/anon4.C (test): Use (double) 3.14 instead of 3.14 in
	comparisons.

+++ gcc/testsuite/g++.dg/reflect/anon4.C
@@ -28,6 +28,6 @@ void test ()
 
   constexpr foo bar1 { .i = 42, .c = 3.14 };
 
-  static_assert (bar1.c == 3.14);
-  static_assert (bar1.[: ^^foo::c :] == 3.14);
+  static_assert (bar1.c == (double) 3.14);
+  static_assert (bar1.[: ^^foo::c :] == (double) 3.14);
 }
